+++ jest.config.cjs
@@ -1,0 +1,10 @@
+/** @type {import('jest').Config} */
+const config = {
+	"transform": {
+		"^.+\\.jsx?$": "babel-jest",
+		"^.+\\.(md|mdx)$": "jest-transformer-mdx",
+	},
+	"transformIgnorePatterns": [],
+};
+
+module.exports = config;