+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.8.4
+## Code Changes
+- refactor code for testability and less mutability
+
+# 0.8.3
+## Code Changes
+- setup Jest for testing
+
 # 0.8.0
 ## Breaking Changes
 - sketches now use deflate compression. Old sketches will break when shared.